Fix indefinitely stuck FPM status processes
After switching to dedicated status port provided by php-fpm, the fpm exporter seems to be stuck on connecting to the fcgi. Only way to fix it is connect another procces to the same fpm port. This suggest some issues in the php (8.0) itself, but to overcome that let's allow next request to acctually try connect and unlock all stuck processes